# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: Toddle Assignment
  description: 'CRUD Apis of Users and Surveys'
  contact: {}
host: '127.0.0.1:3000'
basePath: /v1
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /surveys:
    get:
      tags:
      - "Surveys"
      description: 'Get survey questions details using Id'
      summary: 'http://127.0.0.1:3000/v1/surveys'
      operationId: 'getSurvey'
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: false
          type: string
          description: ''
          default: ''
        - name: id
          in: query
          type: string
          description: 'survey id'
      responses:
        '200':
          description: ''
          schema:
            $ref: "#/definitions/getSurveyResponse"
        '400':
          description: 'Bad request'
          schema:
            $ref: "#/definitions/errorMessage"
        '404':
          description: 'User not found'
        '500':
          description: 'Server error'
    post:
      tags:
      - "Surveys"
      description: 'create a Survey'
      summary: 'http://127.0.0.1:3000/v1/surveys'
      operationId: 'createSurvey'
      parameters:
        - name: Authorization
          in: header
          required: false
          type: string
          description: ''
          default: ''
        - name: Content-Type
          in: header
          required: false
          type: string
          description: ''
          default: 'application/json'
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/surveyMetadata"
      responses:
        '200':
          description: ''
          schema:
            $ref: "#/definitions/surveyCreateResponse"
        '400':
          description: 'Bad request'
          schema:
            $ref: "#/definitions/errorMessage"
        '500':
          description: 'Server error'
...
definitions:
  getSurveyResponse:
    type: "object"
    properties:
      totalQuestions:
        type: "integer"
      questions:
        type: "array"
        items:
          $ref: '#/definitions/surveyQuestion'
...
  surveyQuestion:
    type: "object"
    required:
    - "question"
    - "type"
    properties:
      surveyId:
        type: "string"
      question:
        type: "string"
      type:
        type: "string"
      options:
        type: "array"
        items:
          type: "string"
      canBeSkipped:
        type: "boolean"
  errorMessage:
    type: "object"
    properties:
      message:
        type: "string"
...
  surveyCreateResponse:
    type: "object"
    properties:
      message:
        type: "string"
      id:
        type: "string"
...
  surveyMetadata:
    type: "object"
    properties:
      name:
        type: "string"
      category:
        type: "string"
      subject:
        type: "string"
      topic:
        type: "string"
...